move form logic to useClientContactForm.ts
